feat: Move Update section above Updated
Before, the list of plugins that were just updated would appear above
the list of plugins that are still to be updated. So if you are trying
to update each plugin one at a time, you would have to keep going down
farther and farther to get back to the list of pending updates.

By switching the order of the two sections, the list of pending updates
stays first and makes it easier to read the changes of each plugin
before updating them.