feat(loaders): Allow alt loaders like fnl or tl
I'm playing with https://github.com/folke/lazy.nvim/issues/978

So here's the thing. This is working for me, but I'm not 100% sure of
why it's working. It's not hitting *most* of the code I touched since I
didn't modify anything to touch that code. It is my suspicion that what
I've done is make lazy.nvim more flexible at the cost of breaking
LazyVim, which I don't use. So now I think the best course of action is
to pull LazyVim and configure it to use this alternate loaders code to
see how it falls over