Add extra normalization check
This extra check will more optimistically try to normalize a spec
formatted like this:

```lua
return {
  {'foo/bar'},
  baz = "asdf"
}
```

Before, lazy.nvim would see that it's a table with a length of 1 and
that it's not a list, so it tries to add the table containing the spec
directly, instead of the spec itself. This check makes lazy.nvim proceed
to normalize {'foo/bar'} instead.