fix(lockfile): ensure newline at EOF for lockfile
Text files should end with a newline. This also lets you manually edit
the lockfile in vim without 'fixeol' creating a spurious change for the
added newline.

This change however will create a change in users' lockfiles adding a
newline upon updating, but since the lockfile would be changing anyways
to update lazy.nvim itself, this is likely acceptable.