docs(commands): fix command ordering for sync (#1661)
## Description

After a discussion on Slack we noticed that the description of the sync
command is a bit vague. Some people, including myself, assumed
`clean`/`install`/`update` referred to build steps being performed per
plugin.

Another person mentioned they thought it referred to the Lazy commands,
which does make more sense.
They also noticed that the order of the commands do not match the source
code.

So this PR corrects the order, assuming it was meant to be ordered, and
mentions that those are commands for clarity.