fix: prevent changing buffer in lazy window
Adds the new `winfixbuf` to the window options of the lazy window. This makes the buffer "sticky", preventing accidentally switching to another buffer while in that window, e.g. via `:edit`. See the PR at nvim core https://github.com/neovim/neovim/issues/12517

`winfixbuf` is only available on newer nvim versions, so adding the option is wrapped in a condition checking whether it is supported.