fix(plugin): merge cond property of import specs
When a spec had an `import` property, the `cond` property was treated
the same as `enabled`; any specs to be imported were ignored and treated
as uninstalled if `cond` evaluated to false. This is inconsistent with
how `cond` behaves in a plugin spec without an `import` property.

To make the behavior consistent when using `import` to manage submodules
of plugin specs, merge the `cond` property of the "parent" spec(s)
containing an `import` property with any `cond` property found on the
imported "child" spec(s), including nested `import` specs. The parent
and child `cond` properties are evaluated and combined with boolean
`and`.

Signed-off-by: Ross Williams <ross@ross-williams.net>