fix(rockspec): Check for Lua 5.1 header files.
Previously, rockspec.lua checked if Lua 5.1 was installed to determine
if LuaRocks could build packages for Lua 5.1. This was not sufficient
since the Lua version does not matter as long as the development headers
for Lua 5.1 are available. So Lua 5.1 could be installed and the
LuaRocks packages could still fail to install.

Note: this also checks if a suitable version of Lua is installed since
      the command will fail if Lua is not installed at all.